+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### ✨ Features
-- None
+- [Config] 新增了 `show_discord_avatar` 用于控制转发消息时是否携带用户的 Discord 头像（默认开启）
+- [Config] 新增了 `sync_edit_delete` 用于控制是否同步 Discord 侧编辑 / 删除消息事件到 QQ 侧（默认关闭）
 
 ### 🐞 Bug Fixes
-- [Discord -> QQ] 修复了 Discord 发送链接时有时候链接的参数无法正常显示的 Bug
+- None